pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Guy Torbet", email = "guy@torbet.co" }
]
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.17.2",
    "asyncpg>=0.31.0",
    "fastapi>=0.122.0",
    "pwdlib[argon2]>=0.3.0",
    "pydantic-settings>=2.12.0",
    "pydantic[email]>=2.12.5",
    "pyjwt>=2.10.1",
    "sqlalchemy[asyncio]>=2.0.44",
    "uvicorn>=0.38.0",
]

[project.scripts]
start = "app:start"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
